refactor(solver): extract the sweep reader from the PV audit script
A second instrument needs to read the same layers, and a shared reader keeps the two agreeing about what a layer is. Moved verbatim from scripts/exp002_pv_audit.py, with release() carrying the docstring for the run that proved eviction is not the same as freeing.

No behaviour change; the audit's tests cover release() through the new class.
